fix: recv_fifo for asyncfl (#449)
recv_fifo() takes first_k as one of its arguments to determine the
first_k number of messages are received and returned.

To implement the functionality, asynchronous call
_streamer_for_recv_fifo() was implemented. However the function has a
bug that is triggered in asyncfl. The function adds the first_k
messages into an rx queue and re-adds the remaining messages to the
head of each corresponding end id's rx queue. The remaining messages
then processed in the next recv_fifo call. However,
since _streamer_for_recv_fifo() is an asynchronous call, there is no
guarantee that the messages saved in the previous call will be
processed first. This causes some trainers blocked and made them fail
to participate future rounds.

The issue is fixed as follows. Since recv_fifo() always returns
first_k number of messages from the rx queue, there is no harm even if
the remaining messages are added to the queue. By removing all the
code to handle the remaining code, the code becomes now clean and easy
to understand.